# docs/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: SendaGo Payment API
  version: "1.0"
  description: >
    Fase 1: Payment Link, Invoice, QRIS (as a checkout payment method), Webhook,
    Dashboard. All requests go through api-gateway. Dashboard-auth endpoints use a JWT
    from /auth/login or /auth/signup; the public integration surface uses a merchant
    API key + X-Environment header.
servers:
  - url: http://localhost:8000
components:
  securitySchemes:
    jwt: { type: http, scheme: bearer, bearerFormat: JWT }
    apiKey: { type: http, scheme: bearer, description: "sg_sandbox_... or sg_production_..." }
  parameters:
    XEnvironment:
      name: X-Environment
      in: header
      required: true
      schema: { type: string, enum: [sandbox, production] }
paths:
...
  /payments/{id}:
    get:
      summary: Get transaction status
      security: [{ apiKey: [] }]
      parameters: [{ $ref: "#/components/parameters/XEnvironment" }]
      responses: { "200": { description: OK } }
...